Visualize3D needs to track paths better
- split the module up into different files
- path planner object tracks a traveler and its obstacles, etc.
- rename things accordingly
