+++ acq4/modules/Visualize3D/__init__.py
@@ -1,452 +1,3 @@
-import os
-import queue
-import time
-from threading import Thread
-from typing import Union
-
-import numpy as np
-
-from acq4.devices.Device import Device
-from acq4.devices.OptomechDevice import OptomechDevice
-from acq4.modules.Module import Module
-from acq4.util import Qt
-from acq4.util.future import future_wrap
-from acq4.util.geometry import Plane, Volume, Geometry
-from acq4.util.threadrun import runInGuiThread
-from coorx import Transform, TTransform
-from pyqtgraph import opengl as gl, debug
-from pyqtgraph.opengl.GLGraphicsItem import GLGraphicsItem
-
-
-class Visualize3D(Module):
-    moduleDisplayName = "3D Visualization"
-    moduleCategory = "Utilities"
-
-    win = None
-
-    @classmethod
-    def openWindow(cls):
-        if cls.win is None:
-            cls.win = VisualizerWindow()
-        cls.win.show()
-        cls.win.clear()
-
-    def __init__(self, manager, name: str, config: dict):
-        super().__init__(manager, name, config)
-        runInGuiThread(self.openWindow)
-        for dev in manager.listInterfaces("OptomechDevice"):
-            dev = manager.getDevice(dev)
-            self.win.addDevice(dev)
-
-
-class VisualizerWindow(Qt.QMainWindow):
-    pathStartSignal = Qt.pyqtSignal(object, object, list)
-    newObstacleSignal = Qt.pyqtSignal(str, object, object, object, object)
-    newDeviceSignal = Qt.pyqtSignal(object)
-    pathUpdateSignal = Qt.pyqtSignal(object)
-    focusEvent = Qt.pyqtSignal()
-    _obstacleCache = {}
-    _voxelsCache = {}
-
-    def __init__(self, testing=False):
-        super().__init__(None)
-        self._testing = testing
-        self.setWindowTitle("3D Visualization of all Optomech Devices")
-        self.setGeometry(50, 50, 1000, 600)
-        self.setWindowIcon(Qt.QIcon(os.path.join(os.path.dirname(__file__), "icons.svg")))
-
-        # Main layout is a horizontal splitter
-        self.splitter = Qt.QSplitter(Qt.Qt.Horizontal)
-        self.setCentralWidget(self.splitter)
-
-        # Left side: 3D view
-        self.viewWidget = Qt.QWidget()
-        self.viewLayout = Qt.QVBoxLayout()
-        self.viewWidget.setLayout(self.viewLayout)
-        self.splitter.addWidget(self.viewWidget)
-
-        # Create 3D view
-        self.view = gl.GLViewWidget()
-        self.viewLayout.addWidget(self.view)
-        self.view.setCameraPosition(distance=0.2, azimuth=-90, elevation=30)
-        grid = gl.GLGridItem()
-        grid.scale(0.001, 0.001, 0.001)
-        self.view.addItem(grid)
-        axes = gl.GLAxisItem()
-        axes.setSize(0.1, 0.1, 0.1)
-        self.view.addItem(axes)
-
-        # Right side: Tree widget for device controls
-        self.controlWidget = Qt.QWidget()
-        self.controlLayout = Qt.QVBoxLayout()
-        self.controlWidget.setLayout(self.controlLayout)
-        self.splitter.addWidget(self.controlWidget)
-
-        # Create tree widget for hierarchical device display
-        self.deviceTree = Qt.QTreeWidget()
-        self.deviceTree.setHeaderLabel("Devices")
-        self.deviceTree.setMinimumWidth(250)
-        self.deviceTree.itemChanged.connect(self.handleTreeItemChanged)
-        self.controlLayout.addWidget(self.deviceTree)
-
-        # Path planning controls
-        self.pathPlanToggler = Qt.QCheckBox("Show Path Plan")
-        self.pathPlanToggler.setEnabled(False)
-        self.pathPlanToggler.stateChanged.connect(self.togglePathPlan)
-        self.controlLayout.addWidget(self.pathPlanToggler)
-
-        # Set splitter sizes
-        self.splitter.setSizes([700, 300])
-
-        # Connect signals
-        self.pathStartSignal.connect(self._startPath)
-        self.newObstacleSignal.connect(self._addObstacleVolumeOutline)
-        self.newDeviceSignal.connect(self._addDevice)
-        self.pathUpdateSignal.connect(self._appendPath)
-        self.focusEvent.connect(self._focus)
-
-        # Path planning thread
-        self._pathUpdates = queue.Queue()
-        self._pathWatcherThread = Thread(target=self._watchForPathUpdates)
-        self._pathWatcherThread.daemon = True
-        self._pathWatcherThread.start()
-
-        self._itemsByDevice = {}  # Maps devices to their visualization components
-
-        self._startTarget = gl.GLScatterPlotItem(pos=[], color=(0, 0, 1, 1), size=10, pxMode=True)
-        self._startTarget.setVisible(False)
-        self.view.addItem(self._startTarget)
-
-        self._destTarget = gl.GLScatterPlotItem(pos=[], color=(0, 1, 0, 1), size=10, pxMode=True)
-        self._destTarget.setVisible(False)
-        self.view.addItem(self._destTarget)
-
-        self._activePath = gl.GLLinePlotItem(color=(0.1, 1, 0.7, 0.5), width=1)
-        self._activePath.setVisible(False)
-        self.view.addItem(self._activePath)
-
-        self._previousPath = gl.GLLinePlotItem(color=(1, 0.7, 0, 0.01), width=1)
-        self._previousPath.setVisible(False)
-        self.view.addItem(self._previousPath)
-
-    def clear(self):
-        for dev in list(self._itemsByDevice.keys()):
-            self._removeDevice(dev)
-        self._itemsByDevice = {}
-        self.removePath()
-
-    def focus(self):
-        self.focusEvent.emit()
-
-    def _focus(self):
-        self.show()
-        self.activateWindow()
-        self.raise_()
-
-    def addDevice(self, dev: OptomechDevice):
-        self.newDeviceSignal.emit(dev)
-
-    def _addDevice(self, dev: Union[Device, OptomechDevice]):
-        dev.sigGeometryChanged.connect(self.handleGeometryChange)
-        if (geom := dev.getGeometry()) is None:
-            return
-
-        # Create device entry in data structure
-        self._itemsByDevice[dev] = {"checkboxes": {}}
-
-        # Add geometry to the scene
-        self.addGeometry(geom, dev)
-        dev.sigGlobalTransformChanged.connect(self.handleTransformUpdate)
-        self.handleTransformUpdate(dev, dev)
-
-        # Create tree item for this device
-        deviceItem = Qt.QTreeWidgetItem(self.deviceTree)
-        deviceItem.setText(0, dev.name())
-        deviceItem.setFlags(deviceItem.flags() | Qt.Qt.ItemIsUserCheckable)
-        deviceItem.setCheckState(0, Qt.Qt.Checked)
-        deviceItem.setData(0, Qt.Qt.UserRole, dev)
-        self._itemsByDevice[dev]["tree item"] = deviceItem
-
-        # Create geometry sub-item
-        geomItem = Qt.QTreeWidgetItem(deviceItem)
-        geomItem.setText(0, "Geometry")
-        geomItem.setFlags(geomItem.flags() | Qt.Qt.ItemIsUserCheckable)
-        geomItem.setCheckState(0, Qt.Qt.Checked)
-        geomItem.setData(0, Qt.Qt.UserRole, "geometry")
-        self._itemsByDevice[dev]["checkboxes"]["geometry"] = geomItem
-
-        # Add boundaries if available
-        if bounds := dev.getBoundaries():
-            self._itemsByDevice[dev]["limits"] = {}
-            self.addBounds(bounds, self._itemsByDevice[dev]["limits"])
-
-            # Create limits sub-item
-            limitsItem = Qt.QTreeWidgetItem(deviceItem)
-            limitsItem.setText(0, "Range of Motion")
-            limitsItem.setFlags(limitsItem.flags() | Qt.Qt.ItemIsUserCheckable)
-            limitsItem.setCheckState(0, Qt.Qt.Unchecked)
-            limitsItem.setData(0, Qt.Qt.UserRole, "limits")
-            self._itemsByDevice[dev]["checkboxes"]["limits"] = limitsItem
-
-            # Initially hide limits
-            self.toggleDeviceLimitsVisibility(False, dev.name())
-
-            if hasattr(dev, "sigCalibrationChanged"):
-                dev.sigCalibrationChanged.connect(self.handleGeometryChange)
-
-        # Create obstacle sub-item (initially disabled)
-        obstacleItem = Qt.QTreeWidgetItem(deviceItem)
-        obstacleItem.setText(0, "Obstacle")
-        obstacleItem.setFlags(obstacleItem.flags() | Qt.Qt.ItemIsUserCheckable)
-        obstacleItem.setCheckState(0, Qt.Qt.Checked)
-        obstacleItem.setDisabled(True)  # Disabled until path planning starts
-        obstacleItem.setData(0, Qt.Qt.UserRole, "obstacle")
-        self._itemsByDevice[dev]["checkboxes"]["obstacle"] = obstacleItem
-
-        # Create raw voxels sub-item (initially disabled)
-        voxelsItem = Qt.QTreeWidgetItem(deviceItem)
-        voxelsItem.setText(0, "Raw Obstacle Voxels")
-        voxelsItem.setFlags(voxelsItem.flags() | Qt.Qt.ItemIsUserCheckable)
-        voxelsItem.setCheckState(0, Qt.Qt.Unchecked)
-        voxelsItem.setDisabled(True)  # Disabled until path planning starts
-        voxelsItem.setData(0, Qt.Qt.UserRole, "voxels")
-        self._itemsByDevice[dev]["checkboxes"]["voxels"] = voxelsItem
-
-    def handleTreeItemChanged(self, item: Qt.QTreeWidgetItem, column):
-        parentItem = item.parent()
-        visible = item.checkState(0) == Qt.Qt.Checked
-
-        # A top-level device item
-        if parentItem is None:
-            dev = item.data(0, Qt.Qt.UserRole)
-
-            if dev not in self._itemsByDevice:
-                return
-
-            for ch in range(item.childCount()):
-                child = item.child(ch)
-                name = child.data(0, Qt.Qt.UserRole)
-                child.setDisabled(not visible or name not in self._itemsByDevice[dev])
-                if visible:
-                    self.handleTreeItemChanged(child, None)
-
-            if not visible:
-                self.toggleDeviceLimitsVisibility(False, dev.name())
-                for item in self._itemsByDevice[dev].values():
-                    if hasattr(item, "setVisible"):
-                        item.setVisible(False)
-        else:  # This is a component item
-            dev = parentItem.data(0, Qt.Qt.UserRole)
-            componentType = item.data(0, Qt.Qt.UserRole)
-
-            if dev not in self._itemsByDevice:
-                return
-
-            if componentType in self._itemsByDevice[dev]:
-                if componentType == "limits":
-                    self.toggleDeviceLimitsVisibility(visible, dev.name())
-                else:
-                    self._itemsByDevice[dev][componentType].setVisible(visible)
-
-    def addBounds(self, bounds, displayables_container: dict):
-        for a, b in Plane.wireframe(*bounds):
-            for bound in bounds:
-                if not (bound.allows_point(a) and bound.allows_point(b)):
-                    continue
-            if not self._testing and np.linalg.norm(a - b) > 0.1:
-                continue  # ignore bounds that are really far away
-            edge = gl.GLLinePlotItem(pos=np.array([a, b]), color=(1, 0, 0, 0.2), width=4)
-            edge.setVisible(self.shouldShowPath)
-            self.view.addItem(edge)
-            displayables_container[(tuple(a), tuple(b))] = edge
-
-    def toggleDeviceLimitsVisibility(self, visible, dev_name=None):
-        for key, data in self._itemsByDevice.items():
-            if key.name() == dev_name:
-                for edge in data.get("limits", {}).values():
-                    edge.setVisible(visible)
-
-    def addGeometry(self, geom: Geometry, key=None):
-        if key is None:
-            key = geom.name
-        self._itemsByDevice.setdefault(key, {})
-        self._itemsByDevice[key]["geom"] = geom
-        mesh = geom.glMesh()
-        self.view.addItem(mesh)
-        self._itemsByDevice[key]["geometry"] = mesh
-
-    def _removeDevice(self, dev):
-        if dev not in self._itemsByDevice:
-            return
-
-        # Disconnect signals
-        if not self._testing:
-            dev.sigGeometryChanged.disconnect(self.handleGeometryChange)
-            dev.sigGlobalTransformChanged.disconnect(self.handleTransformUpdate)
-
-        # Remove limits from view
-        for edge in self._itemsByDevice[dev].pop("limits", {}).values():
-            self.view.removeItem(edge)
-            edge.deleteLater()
-
-        # Remove tree item
-        if "tree item" in self._itemsByDevice[dev]:
-            tree_item = self._itemsByDevice[dev].pop("tree item")
-            parent = tree_item.parent()
-            parent.removeChild(tree_item)
-
-        # Remove everything else from view
-        for component in self._itemsByDevice[dev]:
-            self.view.removeItem(component)
-
-        del self._itemsByDevice[dev]
-
-    def handleTransformUpdate(self, moved_device: OptomechDevice, cause_device: OptomechDevice):
-        geom = self._itemsByDevice.get(moved_device, {}).get("geom")
-        if geom is None:
-            return
-        xform = moved_device.globalPhysicalTransform() * geom.transform.as_pyqtgraph()
-        self.setMeshTransform(moved_device, xform)
-
-    def setMeshTransform(self, dev, xform):
-        self._itemsByDevice[dev]["geometry"].setTransform(xform)
-
-    def handleGeometryChange(self, dev: OptomechDevice):
-        self._removeDevice(dev)
-        self._itemsByDevice[dev] = {}
-        self.addDevice(dev)
-
-    def startPath(self, start, stop, bounds):
-        self.pathStartSignal.emit(start, stop, bounds)
-
-    def _startPath(self, start, stop, bounds):
-        # Clear any existing path
-        self.removePath()
-
-        # Enable path planning controls
-        if self.shouldShowPath:
-            self.pathPlanToggler.setEnabled(True)
-            self.pathPlanToggler.setChecked(True)
-
-        # Add initial path line
-        self._appendPath([start, stop])
-
-        self._startTarget.setData(pos=np.array([start]))
-        self._startTarget.setVisible(self.shouldShowPath)
-
-        self._destTarget.setData(pos=np.array([stop]))
-        self._destTarget.setVisible(self.shouldShowPath)
-
-        # Add boundary visualization
-        self.addBounds(bounds)
-
-    @future_wrap
-    def addObstacleVolumeOutline(self, name: str, obstacle: Volume, to_global: Transform, _future):
-        self.newObstacleSignal.emit(name, obstacle, to_global, *obstacle.surface_mesh)
-
-    def _addObstacleVolumeOutline(self, name: str, obstacle: Volume, to_global: Transform, verts, faces):
-        profile = debug.Profiler()
-        device = next((dev for dev in self._itemsByDevice if dev.name() in name), None)
-        if device is None:
-            raise ValueError(f"Could not find device associated with '{name}'")
-        general_checkbox = self._itemsByDevice[device]["checkboxes"]["geometry"].parent()
-        generally_visible = general_checkbox.checkState(0) == Qt.Qt.Checked
-
-        # Create mesh for obstacle surface
-        mesh_data = gl.MeshData(vertexes=verts, faces=faces)
-        mesh = gl.GLMeshItem(
-            meshdata=mesh_data, smooth=True, color=(0.1, 0.1, 0.3, 0.25), shader="balloon", glOptions="additive"
-        )
-        cs_name = obstacle.transform.systems[0].name
-        recenter_voxels = TTransform(
-            offset=(0.5, 0.5, 0.5),
-            from_cs=f"[isosurface of {cs_name}]",
-            to_cs=cs_name,
-        )
-        mesh.setTransform((to_global * obstacle.transform * recenter_voxels).as_pyqtgraph())
-        self.view.addItem(mesh)
-        profile.mark("mesh rendered")
-        self._itemsByDevice[device]["obstacle"] = mesh
-        mesh_checkbox = self._itemsByDevice[device]["checkboxes"]["obstacle"]
-        was_disabled = mesh_checkbox.isDisabled()
-        mesh_checkbox.setDisabled(not generally_visible)
-        if generally_visible and was_disabled:
-            mesh.setVisible(self.shouldShowPath)
-            mesh_checkbox.setCheckState(0, Qt.Qt.Checked)
-        else:
-            mesh.setVisible(self.shouldShowPath and generally_visible and mesh_checkbox.checkState(0) == Qt.Qt.Checked)
-        profile.mark("mesh visible")
-
-        # Create volumetric visualization for raw voxels
-        vol_data = np.zeros(obstacle.volume.T.shape + (4,), dtype=np.ubyte)
-        vol_data[..., :3] = (30, 10, 10)
-        vol_data[..., 3] = obstacle.volume.T * 5
-        vol = gl.GLVolumeItem(vol_data, sliceDensity=10, smooth=False, glOptions="additive")
-        vol.setTransform((to_global * obstacle.transform).as_pyqtgraph())
-        self.view.addItem(vol)
-        profile.mark("voxels rendered")
-        self._itemsByDevice[device]["voxels"] = vol
-        vol_checkbox = self._itemsByDevice[device]["checkboxes"]["voxels"]
-        vol_checkbox.setDisabled(not generally_visible)
-        vol.setVisible(self.shouldShowPath and generally_visible and vol_checkbox.checkState(0) == Qt.Qt.Checked)
-        profile.finish()
-
-    def updatePath(self, path, skip=3):
-        self._pathUpdates.put((path, skip))
-
-    def _watchForPathUpdates(self):
-        n_updates = 0
-        while True:
-            path, skip = self._pathUpdates.get()
-            n_updates += 1
-            if self._testing or n_updates % skip == 0:
-                self.pathUpdateSignal.emit(path)
-                time.sleep(0.02)
-
-    def _appendPath(self, path):
-        if len(self._activePath.pos) > 0:
-            prev = self._activePath.pos
-        else:
-            prev = []
-        if len(self._previousPath.pos) > 0:
-            prev = np.vstack((self._previousPath.pos, prev))
-
-        self._previousPath.setData(pos=prev)
-        self._activePath.setData(pos=np.array(path + path[:-1][::-1]))  # it needs to walk back to the origin
-
-    def togglePathPlan(self, state):
-        visible = state == Qt.QtCore.Qt.Checked
-
-        self._activePath.setVisible(visible)
-        self._previousPath.setVisible(visible)
-        self._startTarget.setVisible(visible)
-        self._destTarget.setVisible(visible)
-
-        for dev, items in self._itemsByDevice.items():
-            generally_visible = items["checkboxes"]["geometry"].parent().checkState(0) == Qt.Qt.Checked
-            if generally_visible:
-                if "obstacle" in items:
-                    obst_visible = items["checkboxes"]["obstacle"].checkState(0) == Qt.Qt.Checked
-                    items["obstacle"].setVisible(visible and obst_visible)
-                if "voxels" in items:
-                    vox_visible = items["checkboxes"]["voxels"].checkState(0) == Qt.Qt.Checked
-                    items["voxels"].setVisible(visible and vox_visible)
-
-    def removePath(self):
-        self._activePath.setVisible(False)
-        self._previousPath.setVisible(False)
-        self._startTarget.setVisible(False)
-        self._destTarget.setVisible(False)
-
-        for items in self._itemsByDevice.values():
-            items["checkboxes"]["obstacle"].setDisabled(True)
-            items["checkboxes"]["voxels"].setDisabled(True)
-            # TODO have these components for each traveler
-            if "obstacle" in items:
-                items["obstacle"].setVisible(False)
-            if "voxels" in items:
-                items["voxels"].setVisible(False)
-
-    @property
-    def shouldShowPath(self):
-        return self.pathPlanToggler.isChecked() or not self.pathPlanToggler.isEnabled()
+from .travelers_proxy import VisualizePathPlan
+from .module import Visualize3D
+from .gui import VisualizerWindow
